+++ Engine/src/Scene/SceneCamera.h
@@ -57,7 +57,7 @@ private:
 
 	float m_PerspectiveNear = 1.0f;
 	float m_PerspectiveFar = -1.0f;
-	float m_Fov = PI * 0.5f;
+	float m_Fov = (float)PI * 0.5f;
 
 	float m_AspectRatio = 16.0f / 9.0f;
 };
